services/ml/premium_service/requirements.txt
Flask==3.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.1
xgboost==2.0.0
requests==2.32.3
psycopg2-binary==2.9.9
